rename and move files in packages
