+++ CurrencyConverter/Java-SpringBoot/src/main/java/edu/campuswien/se/carrental/currencyconverter/SoapServiceApplication.java
@@ -4,9 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ProducingWebServiceApplication {
+public class SoapServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ProducingWebServiceApplication.class, args);
+		SpringApplication.run(SoapServiceApplication.class, args);
 	}
 }
